+++ commercetools.Sdk/commercetools.Sdk.Domain/Carts/LineItemDraft.cs
@@ -10,7 +10,7 @@ namespace commercetools.Sdk.Domain.Carts
 
         public string Sku { get; set; }
 
-        public int Quantity { get; set; }
+        public long Quantity { get; set; }
 
         public Reference<Channel> SupplyChannel { get; set; }
